move out filter state interface
